src/requirements.txt
asgiref==3.3.4
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
Keras==2.3.1
numpy==1.17.3
pandas==1.2.4
SoundFile==0.10.3.post1
joblib==1.0.1
librosa==0.8.0
matplotlib==3.1.1
Pillow==8.2.0
opencv-python==3.4.6.27
tensorflow==2.0.0